moved configuration classes to special package
