+++ src/main/java/org/lekitech/gafalag/configuration/DataV2Configuration.java
@@ -1,4 +1,4 @@
-package org.lekitech.gafalag;
+package org.lekitech.gafalag.configuration;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
